requirements.txt
ipykernel
python-dotenv

streamlit
streamlit-option-menu
streamlit-calendar
google-generativeai
openai

google-api-python-client 
google-auth==2.38.0 
google-auth-oauthlib 
google-auth-httplib2
youtube-transcript-api

schedule

matplotlib
pandas
numpy
seaborn

fastmcp
textblob
gensim